+++ weblate/machinery/deepl.py
@@ -24,6 +24,9 @@ from weblate.machinery.base import MachineTranslation, MissingConfiguration
 DEEPL_TRANSLATE = "{}translate"
 DEEPL_LANGUAGES = "{}languages"
 
+# Extracted from https://www.deepl.com/docs-api/translating-text/response/
+FORMAL_LANGUAGES = {"DE", "FR", "IT", "ES", "NL", "PL", "PT-PT", "PT-BR", "RU"}
+
 
 class DeepLTranslation(MachineTranslation):
     """DeepL (Linguee) machine translation support."""
@@ -52,7 +55,13 @@ class DeepLTranslation(MachineTranslation):
             DEEPL_LANGUAGES.format(settings.MT_DEEPL_API_URL),
             data={"auth_key": settings.MT_DEEPL_KEY},
         )
-        return [x["language"] for x in response.json()]
+        result = {x["language"] for x in response.json()}
+
+        for lang in FORMAL_LANGUAGES:
+            if lang in result:
+                result.add(f"{lang}@FORMAL")
+                result.add(f"{lang}@INFORMAL")
+        return result
 
     def download_translations(
         self,
@@ -65,15 +74,22 @@ class DeepLTranslation(MachineTranslation):
         threshold: int = 75,
     ):
         """Download list of possible translations from a service."""
+        params = {
+            "auth_key": settings.MT_DEEPL_KEY,
+            "text": text,
+            "source_lang": source,
+            "target_lang": language,
+        }
+        if language.endswith("@FORMAL"):
+            params["target_lang"] = language[:-7]
+            params["formality"] = "more"
+        elif language.endswith("@INFORMAL"):
+            params["target_lang"] = language[:-9]
+            params["formality"] = "less"
         response = self.request(
             "post",
             DEEPL_TRANSLATE.format(settings.MT_DEEPL_API_URL),
-            data={
-                "auth_key": settings.MT_DEEPL_KEY,
-                "text": text,
-                "source_lang": source,
-                "target_lang": language,
-            },
+            data=params,
         )
         payload = response.json()